split stream-gallery
